findNearbyIndices() was dramatically rewritten, and verifyRaDec was reintroduced as a valid function.
Added a bunch of default parameters, and some more checks in blindSolve 
to stop the solver from running when it doesn't have enough information


raDec2Xy and xy2RaDec now return PointD's instead of std::pair's because 
Python didn't know how to free the former, and because it's more 
consistent.

I've also added a function so you can specify the parity (whether the 
image is a mirror of the sky or not). If you know this you can half the 
computation time of the code.