Added the function onlyUseBrightestNObjects()
This fixes a misunderstanding of mine about how astrometry.net works. A 
solver_t structure contains the variable endobj, which I thought could 
be used to ignore all but the brightest N objects. Instead it tells the 
solver to give up after trying to match the brightest N objects against 
all objects in the fieldxy structure.

So what I do now is copy the first N objects from the fieldxy into a new 
structure and add that to the solver.