One small change to loadNearbyIndices
When solving a field with a given input position, we only add indices 
near that position to the solver. However, indices that don't cover the 
entire sky, it can happen that no indices are near the requested 
position. If this happens, we now throw an exception, instead of just 
trying (and failing) to solve.