Add MatchMetadata class to help with generation and comparison of metadata.
We now record where we got the data from, and how we used it.  This is
mainly so a human can figure out what catalogues were used, but also allows
some automatic checking.

The MatchMetadata is not propagated to the user, but it could be in the
future.

Bumped the SMATCHV (SourceMatchVector version number) to 2.