Template CreateWcsWithSip on the MatchVector (#2579).
This allows us to fit a Wcs when we have a different match vector than a
ReferenceMatchVector.  This is currently instantiated for ReferenceMatchVector
and SourceMatchVector, which I expect to be the most commonly used.

Tests, example, and usage in Task updated.  Expanded test to check both types.

This is support for #2562 (task to register images).