Attempt to resolve confusion over filter mappings (#2611).
The merge to master of 809662b broke an existing unit test.  Both that test
and the code to read the astrometry.net catalog appear to be confused about the
different mappings.  One mapping (in AstrometryNetDataConfig) configures how
the catalog is constructed (filter name --> catalog column).  The other mapping
(in MeasAstromConfig) configures what filter is returned (filter I have -->
filter I want).

Updated the code to make this cleaner.  Deprecated getCatalogFilterName method
in favour of a new method (getColumnName) that can be applied to both mag
and mag err columns.  Fixed the bad test and added a new one, to ensure both
of the mappings are exercised.