use RAII for index loading/closing (#2879)
When loading indices (when solving, or just reading a catalog)
one can quickly run out of open file descriptors if not careful.
To ensure we don't, use an RAII object ("IndexManager") on the
indices so that files are closed whenever they go out of scope
(exception, looping, etc).

This replaces the fix to addIndices from #2292 with a more
complete solution, and applies it also to getCatalogImpl, which
was discovered to have the same problem in #2879.