fix astrometry_net.i for swig 2.0.11 (#3061)
Was getting a bunch of errors of the form:
    Error: Can't copy typemap (directorout) std::vector< index_t *,std::allocator< index_t * > > = std::vector< index_t *,std::allocator< index_t * > > &DIRECTOROUT

The problem is index_t, which is defined in astrometry.net.
Perhaps something else is defining the same symbol, or swig is
pre-disposed against it, or it's somehow missing the typedef
in astrometry.net.

I tried putting the astrometry.net #includes in a namespace,
but that also imports other useful symbols (e.g., size_t) in
that namespace.  So the simplest solution is to use index_s
instead of index_t (index_t is a typedef of index_s in astrometry.net's
index.h).