astrom: move allFluxes method arguments to config parameter (#3106)
This prevents the need to pass it all over the place, and gives
the (top-level) user more control.  Removed the code triggering
allFluxes on the basis of declared color terms, as that's undue
familiarity with something that should be in a different package.

allFluxes now defaults to True, as I believe this is the usual
desire of the user.