Add wrapper for astrometry.net catalog to cache loading
Loading the entire catalog currently requires reading all the index files,
just to get the metadata (the healpix and nside), which is slow.
Here, the multiindex_t and list of multiindex_t are wrapped by the new
MultiIndexCache and AstrometryNetCatalog classes, respectively.  An
AstrometryNetCatalog can have a cache written for faster loading of the
metadata.  The cache can be generated with a new bin script,
generateANetCache.py.

More could be moved from the LoadAstrometryNetObjectsTask into the
AstrometryNetCatalog class, but leaving that for a more comprehensive
cleanup.