Fix FitTanSipWcsTask to update coords in source catalog
Also update AstrometryTask to take advantage of this change
Also update tests/testFitTanSipWcsTask.py and tests/testAstrometryTask
to test this change, and to report maximum measured sky and pixel error (if a test passes).
Further update tests/testFitTanSipWcsTask to add a test for radial distortion
to test at a few more points and to use a slightly higher order for more tests.

Update testAstrometryTask to check that ref and src cats updated

Updated testAstrometryTask to check that the source catalog's coord fields
and the reference catalog's centroid fields are updated for the new WCS.

Updated testFitTanSipWcsTask to also test radial distortion.

Modified both testAstrometryTask and testFitTanSipWcsTask to report
maximum fit error and to measure x,y pixel error as a radius.