Display usable sources in _matchAndFitWCS
I modified the matcher to return a catalog of usable sources
and modified the debug display code in _matchAndFitWCS to display it.
This change includes a small API change in SourceInfo: I renamed isClean
to isUsable (as I find that name far clearer and it is now far more public
since it is used for returned results).