Overhaul matchOptimisticB's handling of duplicate ref objects
Improved the way matchOptimisticB keeps track of which reference objects
have been used in FinalVerify. Formerly metadata was used in RecordProxy
but this didn't allow saving the closest match. Now a boost multiindex is used
and the match distance is recorded. In theory one can keep the closer match.
In practice this is not being done because it may not be necessary.