Improve matching by culling non-good sources
MatchOptimisticBTask was not culling non-good sources from the match list
before returning. That was a bug I accidentally introduced while making
other changes.
I also changed the definition of good sources as those that are have no
interpolated pixels in their center, as recommended by Paul Price;
this catches cosmic rays and bad pixels, as well as saturation.
The result of re-introducing the culling is that the CFHT demo
not reaches 18.6 mas scatter of bright stars in only 3 iterations
of the outer loop in AstrometryTask, so I reduced the default accordingly
(still not quite the 2 Paul Price requested, but close).