Do not use eups to locate test files
Previously the following occurred:

- Use EUPS to find meas_astrom (find itself)
- Locate astrometry net data in tests/astrometry_net_data subdir
- Instantiate EUPS object using that path
- Setup astrometry_net_data

Now what happens is:

- Locate directory of test using __file__
- Locate astrometry net data in tests/astrometry_net_data subdir
- Set $ASTROMETRY_NET_DATA_DIR to that location

These are equivalent because the EUPS table files in the
local astrometry_net_data subdirs were blank and so there
were no dependencies.

A helper routine has been written to reduce code duplication.
The helper is in a test file of its own and is tested.