Rename ANetAstrometry.astrometry to _astrometry
ANetAstrometry.astrometry is a private method that implements
the core of solving for a WCS (but without handling the
effects of optical distortion, which is done in "solve"
using a distortion context). It is not intended to be
called directly, so I renamed it to _astrometry.
The equivalent public method is "solve".