Explicitly declare test subclasses to aid automated discovery
unittest automated discovery bypasses the suite() function
in the tests and looks for tests in classes that inherit from
unittest.TestCase and start with "test". To enable this the
code is now configured with explicit test classes that
inherit from the base test implementation and also separately
inherit from a class that has the test code. This prevents
the base class tests from being run by mistake.