Simplify readMatches function API
The old API had the reader sending in a butler and dataId, and
optionally: source and match catalog names, an astrometry config,
and source flags.  Most of this is unecessary as the user can
simply pass in the source and match catalogs directly.  The
function can also create the ANetBasicAstrometryTask itself.