Move joinMatchListWithCatalog to setMatchDistance.py
This function used to reside in anetBasicAstrometry.py, but it is
of use to any AstrometryTask so we move it to a seperate file
(currently called setMatchDistance.py, but may rename to matchUtils.py
for inclusion of any functions providing general utilities for match
catalogs).  In addition to the lists of packed matches and sources,
the caller may also provide their initialized AstrometryTask.  If
None is provided, meas.astrom.astrometry.AstrometryTask will be
initialized.