Match edge sources, then reject before fitting
Change the code to allow sources that are are near the edge to be
matched, then rejected before the matches are used for fitting.
Sources that are near the edge have mediocre centroids, so they
should not be used to fit data, but the centroids are typically
good enough to permit matching. This change avoids the situation
of having a false match to a nearby source -- something we have
seen happen in tests. The net result, based on validateDrp.py,
is very small, but it helps a little and does no harm.