Add optional border argument to createMatchMetadata
The reference object loader in meas_algorithm's loadReferenceObjects.py
grows the bbox by the config parameter pixelMargin: doc = "Padding to
add to 4 all edges of the bounding box (pixels)". This is set to 50 by
default but is not reflected by the radius parameter set in the metadata,
so some matches may reside outside the circle searched within this radius
when unpersisting the match list.  This change allows for the pixelMargin
to be reflected in the radius set in the metadata by adding an optional
argument to specify by how many pixels the bbox should be grown before
determining the radius.