createMatchMetadata: add additional interface
The 'match metadata' stores data we need to regenerate the matches
(i.e., cone search parameters). The current interface uses a bbox
and Wcs, and converts those to the cone search parameters that are
stored in the metadata; but there has not been an interface to
construct the 'match metadata' given the cone search parameters.

Created a new class, MatchMetadata for the purpose of simplifying
the creation of the 'match metadata'. Chose to use a class rather
than a new factory function because it's more fundamental. The old
interface now uses the new class.