add task to perform direct matching to a reference catalog
The two AstrometryTasks (MatchOptimisticBTask and ANetAstrometryTask)
support matching, but these are optimised for determining astrometric
solutions so there are a variety of contortions involved, including
allowing rotations and distortions and filtering of the input catalog.
This new DirectMatchTask is much simpler:
* no exposure required, only a catalog;
* no rotations or translations are permitted, only a straight match;
* no filtering of the input catalog, so all matches are returned; and
* no exception is raised when there are no matches.
This makes it suitable for gathering a list of matches for QA purposes.