Use new accessors in afw to improve test coverage.
This also provides the only test coverage for the accessors themselves,
which would ideally be tested in afw as well.  However, it's much
easier to test them here (afw's tests seem to currently lack any way
to mock up a TanWcs easily), and DM-9901 may result in moving this
entire test to afw anyway.