Created MatchTolerance class.
astrometry.py: Changed toleranceStruct to an instance of a MatchTolerance class.
matchOptimisticBContinued.py: Defined MatchTolerance class for this matcher and replaced toleranceStruct with this class.
matchPessimisticB.py: Defined MatchTolerance class for this matcher and replaced toleranceStruct with this class.

matchPessimisticB.py: Changed patternSkipList to failedPatternList
ref_match.py: Removed toleranceStruct and replaced it with match_tolerance.