reduce namespace confusion for matchOptimisticB
The 'matchOptimisticB' namespace is overloaded, which makes it
impossible to do:
    import lsst.meas.astrom.matchOptimisticB.matchOptimisticBContinued
    lsst.meas.astrom.matchOptimisticB.matchOptimisticBContinued.MatchOptimisticBTask
results in:
    AttributeError: 'builtin_function_or_method' object has no attribute 'matchOptimisticBContinued'

The 'matchOptimisticB' module contains a function called
'matchOptimisticB' which was being imported over the top of the
module, so leave it in the module. Move the MatchOptimisticBTask
from the 'matchOptimisticB' module so it can be accessed
without problems. Added a test that detects the problem.