Add maxMeanDistanceArcsec config to impose WCS fail
It is desirable to have a (configurable) threshold above which even a
converged WCS fit will be deemed a failure to preclude it from
contaminating downstream processing (namely, a calexp with a horrendous
WCS fit making its way into a coadd).

The value to be thresholded against is the mean on-sky distance of the
astrometric fit, the same one as (would) be persisted in our detector/visit
summary catalogs.

The appropriate value for this threshold will depend on the data and
processing workflow.  For example, if the single frame measurement WCS is
the final WCS to be associated with an exposure before moving on to the
coaddition stage, one might want to be fairly strict with this threshold
so as not to allow exposures with "converged" but "bad" WCS fits making
their way into a coadd.  A rough upper bound on this threshold of about
half a pixel has been suggested in this case.  On the other hand, if the
workflow includes an external (and global) astrometric fitting step (e.g.
jointcal) , then it may be possible to loosen this constraint in the hopes
of allowing the external calibration to recover a good WCS fit for a "bad"
initial fit from single frame processing.

Since the main Rubin/LSST processing workflow will indeed include an
external/global calibration step, we set the default here to match the
maximum possible "badness" of the SFM WCS that it could conceivably
recover from, which is currently on the order of 0.5 arcsec.