Add option to cull refcat based on fiducial zeropoints
The loaded reference catalogs can potentially have a magnitude range
very different from the source catalog to which it's being matched.
This situation is both wasteful and potentially dangerous (as having
many objects in the reference catalog that are not in the source
catalog can lead to bad fits).  The fiducial zeropoints give us a
decent estimate of the exposure's zeropoint (which we have not
measured at this stage of the calibration), so use these (with
padding to account for uncertainty in the fiducials) to cul the
reference catalog to a magnitude range that roughly matches that
of the source catalog.