Update fiducial-based reference object culling
The reference selector in play for astrometry uses the value in
anyFilterMapsToThis for its flux reference (appropriately as some
bands have very few good flux measurements, particularly in crowded
fields, and only the centroid is required for the astrometric fit).
The default in pipe_task's calibrateImage is currently set to the
Gaia "phot_g_mean" flux (which is actually closest to r band).
Since the average color of the reference stars can be quite large,
this means the culling based on flux could be biased away from the
typical reference object colors.  Here we attempt to account for
this shift by computing the mean color of the loaded reference
catalog, where possible (i.e. there are enough reference sources
with well-measured flux in the band of the observation).

The most likely scenario for non-typical conditions will be an
effective exctinction, so we lessen the culling at the bright end
of the reference catalog.